backend/requirements.txt
fastapi
uvicorn
openai
pymongo
python-dotenv
python-multipart
